# docs/openapi.yaml
openapi: 3.0.3
info:
  title: My Interview API
  version: 1.0.0
  description: Frontend contract verified against the current Spring Boot implementation.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/interview-templates:
    get:
      tags: [JD and Template]
      operationId: listInterviewTemplates
      summary: List owned or public templates
      parameters:
        - name: scope
          in: query
          schema: {type: string, enum: [mine, public], default: mine}
        - name: page
          in: query
          schema: {type: integer, minimum: 0, default: 0}
        - name: size
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
      responses:
        '200':
          description: Template page.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TemplatePage'}
        '400': {$ref: '#/components/responses/BadRequest'}
  /api/interview-templates/{templateId}:
    parameters:
      - {$ref: '#/components/parameters/TemplateId'}
    get:
      tags: [JD and Template]
      operationId: getInterviewTemplate
      summary: Get an interview template
      responses:
        '200':
          description: Owned or public template.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InterviewTemplate'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      tags: [JD and Template]
      operationId: updateInterviewTemplate
      summary: Update a draft interview template
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateInterviewTemplateRequest'}
      responses:
        '200':
          description: Updated draft template.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InterviewTemplate'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
  /api/interview-templates/{templateId}/confirm:
    parameters:
      - {$ref: '#/components/parameters/TemplateId'}
    post:
      tags: [JD and Template]
      operationId: confirmInterviewTemplate
      summary: Confirm and freeze a template
      requestBody: {$ref: '#/components/requestBodies/TemplateVersion'}
      responses:
        '200':
          description: Confirmed and frozen template.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InterviewTemplate'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/interview-templates/{templateId}/publish:
    parameters:
      - {$ref: '#/components/parameters/TemplateId'}
    post:
      tags: [JD and Template]
      operationId: publishInterviewTemplate
      summary: Publish a confirmed template
      description: Requires ADMIN and ownership.
      requestBody: {$ref: '#/components/requestBodies/TemplateVersion'}
      responses:
        '200':
          description: Published template.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InterviewTemplate'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/interview-templates/{templateId}/unpublish:
    parameters:
      - {$ref: '#/components/parameters/TemplateId'}
    post:
      tags: [JD and Template]
      operationId: unpublishInterviewTemplate
      summary: Unpublish a template
      description: Requires ADMIN and ownership.
      requestBody: {$ref: '#/components/requestBodies/TemplateVersion'}
      responses:
        '200':
          description: Unpublished template.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InterviewTemplate'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/interview-templates/{templateId}/archive:
    parameters:
      - {$ref: '#/components/parameters/TemplateId'}
    post:
      tags: [JD and Template]
      operationId: archiveInterviewTemplate
      summary: Archive a template
      requestBody: {$ref: '#/components/requestBodies/TemplateVersion'}
      responses:
        '200':
          description: Archived and unpublished template.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InterviewTemplate'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
  parameters:
...
    TemplateId:
      name: templateId
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
  requestBodies:
    TemplateVersion:
      required: true
      content:
        application/json:
          schema: {$ref: '#/components/schemas/TemplateVersionRequest'}
  responses:
    BadRequest:
      description: Invalid input.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
...
    Forbidden:
      description: Authenticated but not allowed.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    NotFound:
      description: Resource not found or inaccessible.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    Conflict:
      description: State, idempotency, version, or limit conflict.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
...
    Unprocessable:
      description: Semantically invalid content or failed processing.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
...
  schemas:
    ApiError:
      type: object
      required: [timestamp, status, code, message, path]
      properties:
        timestamp: {type: string, format: date-time}
        status: {type: integer}
        code: {type: string}
        message: {type: string}
        path: {type: string}
        fieldErrors:
          type: object
          nullable: true
          additionalProperties: {type: string}
...
    JobAnalysis:
      type: object
      required: [sufficientJobContext, sourceLanguage, summary, keySkills]
      properties:
        sufficientJobContext: {type: boolean}
        sourceLanguage: {type: string, minLength: 1, maxLength: 20}
        jobTitle: {type: string, maxLength: 150, nullable: true}
        targetSeniority: {type: string, maxLength: 100, nullable: true}
        domain: {type: string, maxLength: 150, nullable: true}
        summary: {type: string, minLength: 1, maxLength: 3000}
        keySkills:
          type: array
          minItems: 1
          maxItems: 30
          items:
            type: object
            required: [name, level]
            properties:
              name: {type: string, minLength: 1, maxLength: 150}
              level: {type: string, enum: [MUST_HAVE, NICE_TO_HAVE]}
              description: {type: string, maxLength: 1000, nullable: true}
...
    InterviewTemplateSummary:
      type: object
      required: [id, sourceJobDescriptionId, title, confirmed, published, updatedAt]
      properties:
        id: {type: integer, format: int64}
        sourceJobDescriptionId: {type: integer, format: int64}
        title: {type: string}
        jobTitle: {type: string, nullable: true}
        targetSeniority: {type: string, nullable: true}
        confirmed: {type: boolean}
        published: {type: boolean}
        archivedAt: {type: string, format: date-time, nullable: true}
        updatedAt: {type: string, format: date-time}
    InterviewTemplate:
      allOf:
        - {$ref: '#/components/schemas/InterviewTemplateSummary'}
        - type: object
          required: [content, version, createdAt]
          properties:
            content: {$ref: '#/components/schemas/JobAnalysis'}
            confirmedAt: {type: string, format: date-time, nullable: true}
            publishedAt: {type: string, format: date-time, nullable: true}
            version: {type: integer, format: int64}
            createdAt: {type: string, format: date-time}
    TemplatePage:
      type: object
      required: [content, page, size, totalElements]
      properties:
        content: {type: array, items: {$ref: '#/components/schemas/InterviewTemplateSummary'}}
        page: {type: integer}
        size: {type: integer}
        totalElements: {type: integer, format: int64}
    UpdateInterviewTemplateRequest:
      type: object
      required: [title, content, expectedVersion]
      properties:
        title: {type: string, minLength: 1, maxLength: 200}
        content: {$ref: '#/components/schemas/JobAnalysis'}
        expectedVersion: {type: integer, format: int64, minimum: 0}
    TemplateVersionRequest:
      type: object
      required: [expectedVersion]
      properties:
        expectedVersion: {type: integer, format: int64, minimum: 0}
...
security:
  - bearerAuth: []